backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
groq==0.31.1  # Changed to latest version
redis==5.0.1
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4